+++ packages/cli/src/cli.ts
@@ -6,7 +6,7 @@ import { gzip } from 'node-gzip';
 import { program } from '@caporal/core';
 import { Logger, NodeIO } from '@gltf-transform/core';
 import { ALL_EXTENSIONS } from '@gltf-transform/extensions';
-import { AOOptions, CenterOptions, DedupOptions, PartitionOptions, ResampleOptions, SequenceOptions, UnweldOptions, WeldOptions, ao, center, dedup, metalRough, partition, resample, sequence, unweld, weld } from '@gltf-transform/lib';
+import { AOOptions, CenterOptions, DedupOptions, PartitionOptions, PruneOptions, ResampleOptions, SequenceOptions, UnweldOptions, WeldOptions, ao, center, dedup, metalRough, partition, prune, resample, sequence, unweld, weld } from '@gltf-transform/lib';
 import { inspect } from './inspect';
 import { DracoCLIOptions, ETC1S_DEFAULTS, Filter, Mode, UASTC_DEFAULTS, draco, merge, toktx, unlit } from './transforms';
 import { Session, formatBytes } from './util';
@@ -394,6 +394,24 @@ compression and instancing, to be more effective.
 	.action(({args, options, logger}) =>
 		Session.create(io, logger, args.input, args.output)
 			.transform(dedup(options as unknown as DedupOptions))
+	);
+
+// PRUNE
+program
+	.command('prune', 'Removes unreferenced properties from the file')
+	.help(`
+Removes properties from the file if they are not referenced by a Scene. Helpful
+when cleaning up after complex workflows or a faulty exporter. This function
+may (conservatively) fail to identify some unused extension properties, such as
+lights, but it will not remove anything that is still in use, even if used by
+an extension. Animations are considered unused if they do not target any nodes
+that are children of a scene.
+	`.trim())
+	.argument('<input>', INPUT_DESC)
+	.argument('<output>', OUTPUT_DESC)
+	.action(({args, options, logger}) =>
+		Session.create(io, logger, args.input, args.output)
+			.transform(prune(options as unknown as PruneOptions))
 	);
 
 // DRACO